+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file. This change
 log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+* SVG diagrams can be resized to fit screen.
 
 ## [2.0.228] - 2021-12-14
 ### Changed